src/main/java/ru/ylab/model/WorkSpace.java
package ru.ylab.model;

public class WorkSpace implements Resource {
    private int id;
    private String name;
    private int capacity;

    public WorkSpace(int id, String name, int capacity) {
        this.id = id;
        this.name = name;
        this.capacity = capacity;
    }

    @Override
    public int getId() {
        return id;
    }

    @Override
    public String getName() {
        return name;
    }

    @Override
    public int getCapacity() {
        return capacity;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setCapacity(int capacity) {
        this.capacity = capacity;
    }
}

